fix: profile change password lint
