+++ client/src/components/profile/constants.js
@@ -1,0 +1,1 @@
+export const DEFAULT_PROFILE_IMAGE = '/default-profile.png';
